+++ modules/http-fauth/src/validator.rs
@@ -80,7 +80,7 @@ impl ConfigurationValidator for ForwardedAuthenticationConfigurationValidator {
                 validate_request_header(auth_to)?;
                 validate_nested!(auth_to, used(sub), last, optional args(1) => [ServerConfigurationValue::Boolean(_, _)] | args(0) => [ServerConfigurationValue::Boolean(_, _)]);
                 validate_nested!(auth_to, used(sub), intercept_errors, optional args(1) => [ServerConfigurationValue::Boolean(_, _)] | args(0) => [ServerConfigurationValue::Boolean(_, _)]);
-                validate_nested!(auth_to, used(sub), no_append_uri, optional args(1) => [ServerConfigurationValue::Boolean(_, _)] | args(0) => [ServerConfigurationValue::Boolean(_, _)]);
+                validate_nested!(auth_to, used(sub), append_uri, optional args(1) => [ServerConfigurationValue::Boolean(_, _)] | args(0) => [ServerConfigurationValue::Boolean(_, _)]);
                 check_unused_subdirectives!(auth_to, sub, &mut ctx.diagnostics, ctx.scope.clone());
             });
         }